requirements.txt
Django==3.1
djangorestframework==3.11.1
numpy==1.18.4
scikit-learn==0.24.2
transformers==4.18.0
captum==0.5.0
pandas==1.2.2
seaborn==0.11.1
torch==1.9.0